refactor: kebab case, also title color
