+++ app/(header-default)/export-sirene/Filters.tsx
@@ -11,10 +11,10 @@ import { codesNAFRev2 } from '#utils/helpers/formatting/metadata/codes-NAF-rev-2
 import { codesSectionNAF } from '#utils/helpers/formatting/metadata/codes-section-NAF';
 import { Dispatch, SetStateAction } from 'react';
 import { getEffectifLabel } from './constants';
-import { ExtendedExportCsvInput } from './ExportCsv';
-import { LocationFilter } from './LocationFilter';
-import LocationTags from './LocationTags';
-import SiretFilter from './SiretFilter';
+import { ExtendedExportCsvInput } from './export-csv';
+import { LocationFilter } from './location-filter';
+import LocationTags from './location-tags';
+import SiretFilter from './siret-filter';
 import styles from './styles.module.css';
 
 export const selectedEffectifLabel = (filters: ExtendedExportCsvInput) => {
