+++ app/(header-default)/export-sirene/export-csv.tsx
@@ -5,9 +5,9 @@ import { formatDate, formatNumber } from '#utils/helpers';
 import { ExportCsvInput } from 'app/api/export-sirene/input-validation';
 import { useState } from 'react';
 import { getEffectifCode } from './constants';
-import Filters from './Filters';
-import FiltersSummary from './FiltersSummary';
-import InfoSection from './InfoSection';
+import Filters from './filters';
+import FiltersSummary from './filters-summary';
+import InfoSection from './info-section';
 import styles from './styles.module.css';
 
 export interface ExtendedExportCsvInput extends ExportCsvInput {
@@ -220,9 +220,7 @@ export default function ExportCsv() {
 
   return !showResults || !countResult ? (
     <div className={styles.exportCsv}>
-      <h1 className={styles.title}>
-        Générez une liste CSV à partir du répertoire Sirene
-      </h1>
+      <h1>Générez une liste CSV à partir du répertoire Sirene</h1>
       <InfoSection />
       <form onSubmit={handleCountSubmit}>
         <Filters filters={filters} setFilters={setFilters} />
