+++ app/(header-default)/export-sirene/filters-summary.tsx
@@ -3,8 +3,8 @@ import { categoriesJuridiquesNiveau1 } from '#utils/helpers/formatting/metadata/
 import { categoriesJuridiquesNiveau2 } from '#utils/helpers/formatting/metadata/categories-juridiques-niveau-2';
 import { codesNAFRev2 } from '#utils/helpers/formatting/metadata/codes-NAF-rev-2';
 import { codesSectionNAF } from '#utils/helpers/formatting/metadata/codes-section-NAF';
-import { ExtendedExportCsvInput } from './ExportCsv';
-import { selectedEffectifLabel } from './Filters';
+import { ExtendedExportCsvInput } from './export-csv';
+import { selectedEffectifLabel } from './filters';
 
 export default function FiltersSummary({
   filters,
